requirements.txt
django>=3.1,<3.2
uwsgi==2.0.19.1
psycopg2==2.8.5
dj-database-url==0.5.0
whitenoise==5.2.0
django-storage-url==0.5.0
boto3==1.14.49
django-pipeline
django-autoslug
awesome-slugify
# unicode-slugify==0.1.3
django-parler
django-ckeditor
django-taggit
django-taggit-autosuggest
pillow
sorl-thumbnail
